requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aioredis==1.2.0           # via channels-redis
amqp==2.5.1               # via kombu
asgiref==3.2.2            # via channels, channels-redis, daphne
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.1      # via aioredis
attrs==19.1.0             # via automat, twisted
autobahn==19.9.3          # via daphne
automat==0.7.0            # via twisted
billiard==3.6.1.0         # via celery
celery==4.3.0
certifi==2019.9.11        # via requests
cffi==1.12.3              # via cryptography
channels-redis==2.4.0
channels==2.2.0
chardet==3.0.4            # via requests
colorama==0.4.1           # via rednose
confusable-homoglyphs==3.2.0  # via django-registration
constantly==15.1.0        # via twisted
cryptography==2.7         # via autobahn
daphne==2.3.0             # via channels
django-appconf==1.0.3     # via django-imagekit
django-bootstrap-pagination==1.7.1
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-nose==1.4.6
django-redis==4.10.0
django-registration==3.0.1
django-static-precompiler==1.8.2
django==1.11.24
entrypoints==0.3          # via flake8
flake8==3.7.8
foursquare==1!2019.9.11
gunicorn==19.9.0
hiredis==1.0.0            # via aioredis
httplib2==0.13.1
hyperlink==19.0.0         # via twisted
idna==2.8                 # via hyperlink, requests, yarl
importlib-metadata==0.23  # via kombu
incremental==17.5.0       # via twisted
isodate==0.6.0
jsonfield==2.0.2
kegbot-api==1.1.0
kegbot-pyutils==0.1.8
kombu==4.6.4              # via celery
mccabe==0.6.1             # via flake8
mock==3.0.5
more-itertools==7.2.0     # via zipp
msgpack==0.6.1            # via channels-redis
multidict==4.5.2          # via yarl
mysqlclient==1.4.4
nose==1.3.7               # via django-nose
oauthlib==3.1.0           # via requests-oauthlib
pilkit==2.0               # via django-imagekit
pillow==6.1.0
protobuf==3.9.1
pycodestyle==2.5.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via flake8
pyhamcrest==1.9.0         # via twisted
pysocks==1.7.0            # via tweepy
python-gflags==3.1.2
pytz==2019.2
pyyaml==5.1.2             # via vcrpy
redis==3.3.8
rednose==1.3.0
requests-mock==1.7.0
requests-oauthlib==1.2.0
requests==2.22.0
six==1.12.0               # via autobahn, automat, cryptography, django-appconf, django-imagekit, foursquare, isodate, mock, protobuf, pyhamcrest, requests-mock, tweepy, txaio, vcrpy
termstyle==0.1.11         # via rednose
tweepy==3.8.0
twisted==19.7.0           # via daphne
txaio==18.8.1             # via autobahn
urllib3==1.25.3           # via requests
vcrpy==2.1.0
vine==1.3.0               # via amqp, celery
whitenoise==4.1.3
wrapt==1.11.2             # via vcrpy
yarl==1.3.0               # via vcrpy
zipp==0.6.0               # via importlib-metadata
zope.interface==4.6.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via protobuf, pyhamcrest, rednose, termstyle, zope.interface
